+++ History.md
@@ -1,9 +1,12 @@
 ## vNEXT
 
+## v0.0.11
+
 * Make `famousCmp.transitions` public so that it's easier to create custom
 named transitions for RenderController.
 * Add `HeaderFooterLayout` view (closes #18)
 * Add `Flipper` view (TODO, ability to create new contexts for perspective)
+* Add `EdgeSwapper` view
 
 ## v0.0.10